fixed servo problem with Timer substituing it with ServoTimer2 modified library

+++ assignment-2/src/smart-coffee-machine/sensors/potentiometer/PotentiometerManager.cpp
@@ -12,6 +12,6 @@ void PotentiometerManager::computeRead() {
     this->potentiometerValue = analogRead(this->pin);
 }
 
-uint8_t PotentiometerManager::getValue() {
+int PotentiometerManager::getValue() {
     return this->potentiometerValue;
 }
